+++ tests/tests_plugin_ods.py
@@ -76,13 +76,27 @@ class PluginOdsTestCase(utils.RowsTestMixIn, unittest.TestCase):
     def test_issue_320_empty_cells(self):
         result = rows.import_from_ods("tests/data/empty-cells.ods")
         header = "f1 f2 f3 f4 f5".split()
-        print(result[0])
         data = [
-            [getattr(result[index], field) for field in header]
-            for index in range(5)
+            [getattr(row, field) for field in header] for row in result
         ]
         assert data[0] == ["r1f1", "r1f2", None, "r1f4", "r1f5"]
         assert data[1] == ["r2f1", None, "r2f3", "r2f4", "r2f5"]
         assert data[2] == [None, "r3f2", "r3f3", "r3f4", "r3f5"]
         assert data[3] == ["r4f1", "r4f2", "r4f3", "r4f4", None]
         assert data[4] == [None, None, "r5f3", "r5f4", "r5f5"]
+
+    def test_cell_range(self):
+        result = rows.import_from_ods(
+            "tests/data/all-field-types.ods",
+            start_row=3,
+            end_row=5,
+            start_column=1,
+            end_column=4,
+        )
+        header = "field_3 field_456 field_456_2 field_12".split()
+        data = [
+            [getattr(row, field) for field in header] for row in result
+        ]
+        assert len(data) == 2
+        assert data[0] == [4, 7.89, 7.89, Decimal("0.1364")]
+        assert data[1] == [5, 9.87, 9.87, Decimal("0.1314")]